Introduce the concept of host harness to mlir-miopen-driver with a test case.
When `--host` is specified along with a filename, `mlir-miopen-driver` would
load it, parse it, and populate 2D convolution logic inside the entry point
function.

An entry point function could be specified with `--entry-point` flag, the
default value is `conv2d`. A `gpu.launch` is expected inside the entry point
function and 2D convolution logic would be populated inside the operation.

An example, `conv2d_harness.mlir` is included in this commit to depict the
usage.

The purpose of this commit is to allow quick prototyping and avoid
time-consuming logic to populate host-side logic.