Improve XDLOPS logic.
Reduce VGPR usage in KReduction case.

Originally a[] is allocated with KPerBlock * MRepeats, but that's more
than what's actually needed. Change to KPerBlock / num_input_blks *
MRepeats in KReduction case. The same goes with b[].

With this change, in KReduction case, we can use vector.transfer_read
and do not have to do strided loads.