Annotate miopen.alloc as an operation that allocates memory (#205)
* Annotate miopen.alloc as an operation that allocates memory

MLIR has an annotation that indicates when an operation allocates
memory, which is present in the GPU dialect. This operation is added
to the miopen.alloc op as well, enabling us to take advantage of
optimization passes that operate on heap buffers before the lowering
to GPU dialect takes place.

The alloc kernel test has been modified to add dummy stores. Not
including these stores caused the memory allocations to be optimized
out.

* Remove stale commented-out include per @whchung