Initial test of the one big Jenkinsfile
This Jenkinsfile is parameterized by "nightly" and "canXdlops" -
booleans that should be set to reflect if the job is being run on pull
requests or not (to remove expensive/unneeded tests) and whether the
job is being run in a CI that can run xdlops (that is, if we're on the
private or public CI).

If this file works as planned, the idea is that there'll be four
Jenkins jobs, two on each CI server, which'll run all the relevant
testing. An advantage to this is that a lot of common infrastructure
is factored out, as in the buildProject() function, and that we can
see all our CI jobs in one place. This also eliminates the need to
have many different CI pipeline configurations, simplifying the
administration of the cluster.