Deduplicate various steps in CI files (#220)
* Deduplicate various steps in CI files

 - Add most of the flags we use in CI to the CMakeLists.txt

 - Introduce a BUILD_FAT_LIBMLIRMIOPEN to control building the fat
MIOpen interface library and avoid building it when it's not needed

 - Move the installation of said library to a separate component, so
 now we don't need to run our build tools with root permissions

 - Replace many shell scripts that call cmake and a build tool with
 calls to the cmake plugin

- Similarly, move PATH changes into the Jenkins configuration

Many of these CI files still contain similar steps (like various
compile/test combinations). Once I have a better sense of what tests
are being run and where, I will be able to combine these files into a
something less redundant and include, for example, conditions for
excluding longer tests on PR-triggered CI.

* Newlines in bad places break everything

* Missed a steps{} directive

* include/includes change

* Image name is wrong

* Call depenndency installer

* Adjust Docker image label

* Update path setup

* Make MIOpen build actually run in parallel

* SourceDir shouldn't in the thing

* Skip default checkout for test nodes

* Move options block to where it should be

* Put back in a config directive the merge threw out.

* Initial test of the one big Jenkinsfile

This Jenkinsfile is parameterized by "nightly" and "canXdlops" -
booleans that should be set to reflect if the job is being run on pull
requests or not (to remove expensive/unneeded tests) and whether the
job is being run in a CI that can run xdlops (that is, if we're on the
private or public CI).

If this file works as planned, the idea is that there'll be four
Jenkins jobs, two on each CI server, which'll run all the relevant
testing. An advantage to this is that a lot of common infrastructure
is factored out, as in the buildProject() function, and that we can
see all our CI jobs in one place. This also eliminates the need to
have many different CI pipeline configurations, simplifying the
administration of the cluster.

* Remove a bunch of withCmake calls that aren't needed

* Missing close brace

* Correct the syntax for sequential-in-parallel

* Clean up some syntax issues, hopefully

* Remove large Jenkinsfile - development moving to a new branch

* Fix test failures, especially when miopen driver isn't built

1. Disable a standalone test that's broken and irrelevant to us
2. Make certain miopen-related directories tests conditional on the
miopen driver being built
3. Disable some MIOpen dialect tests that use the driver when the
driver isn't compiled

* Add in sourceDir for LLVM build

* Maybe it's nested docker() nodes that cause issues - will revert if this isn't it.

* Move scripts into MIOpen test environment

* Switch image to one with 'parallel' in it

* Try out alwaysPull, see what it does