Create one big Jenkinsfile to handle all the CI tasks
This new Jenkinsfile is parameterized by two parameters, nightly (which
controls whether to run the more extensive tests that are part of a
nightly run) and canXdlops (which controls whether the CI is being run
in a cluster that has xdlops support, that is, whether this is on a
public or private CI instance).

As far as I can tell, transitioning to this will involve updating some
of the Jenkins jobs to point to this new `Jenkinsfile` and setting the
parameters appropriately (since, with how I've set this up, parameter
values are sticky).

The advantages to this arrangement are
1. The CI steps are in one place
2. A bunch of duplicate CI work (such as multiple builds of the
library) is removed
3. We can factor out common things, like the name of the docker image,
thus reducing the copy-paste burden.

Note that there is no top level agent or environment block in order to
work around a bug in the Durable Tasks plugin that causes the build to
hang when starting containers from some contexts (like another
container agent or an environment{} block). This does have the
advantage of letting us move jobs that don't need GPUs to nodes that
don't have them.