Add argument validations to MIOpen driver.
Add a general validation that prevents mixed data types.
Also, ensure MIOpen does not try to generate kernels for:
- bf16
- Layouts other than NCHW or NHWC (since they haven't been tested)

Finally, clean up convolution config by removing an unesed variable.

This commit is the MLIR side of a move to move more of the applicability
checks for the MLIR solver to MLIR from MIOpen.