When the host doesn't have a suitable GPU, mark GPU-specific tests as 'unsupported.'
Use rocm_agent_enumerator to see if the host has any useful GPUs.  If
not, then mlir-rocm-runner will barf, so we don't want to run tests that
use it.  Set a config flag, no_AMD_GPU, and use it to disable some
tests, which will then be reported as "unsupported."

Use stdout=subprocess.PIPE instead of capture_output=True for
compatibility with python 3.6.

Rearrange the mlir-miopen-driver tests a bit to reduce the scope of the
"unsupported" test.

Move the rocm_agent_enumerator test from lit.local.cfg to
lit.site.cfg.py.in.  That allows testing in only two places and
simplifies lit.local.cfg to just checking a config flag.