Remove most of the boilerplate of defining ops with assemblyFormat
Instead of defining parsers and printers in MIOpenOps.cpp, we use the
assemblyFormat mechanism in MLIR tablegen to specify the format
declaritively.

This has lead to some breaking changes, mainly around ops that take
variadic arguments, which now can't rely on the implicit equalities
between the variadic input and output we were imposing (adding back
support for that is something we might want to give upstream).
Similarly, the types of variadic arguments are no longer inferred.

This changes the printing format for:
  - move_pos_v2
  - threadwise_copy
  - threadwise_copy_v2
  - blockwise_gemm_v2
  - xdlops_gemm_v2
  - threawise_load
  - threadwise_store
and so the tests have been adjusted accordingly

Similarly, fill now needs to give the type of both its
arguments (supporting that equality would be a nice addition tho MLIR,
but may not be worth the bother if we move to linalg.fill).

While I was accounting for these changes, I also fixed some extremely
long lines in the tests.