[MLIR][GPU] Add gpu.warp_swizzle and its lowering on AMD GPUs.
The gpu.warp_swizzle is a lowest-common-denominator op for moving data
between threads/lanes on a GPU. It allows arbitrary permutations of
data within groups of 4 adjacent threads, since more complex
permutations may be expensive or only available on some platforms.

(The op that takes a computed lane ID, which also would be supported
everywhere, is not one that is added here as it would perform poorly,
likely require synchronization, and is not needed for the applications
that led to devising this intrinsic.)

This commit also adds an intrinsic to the ROCDL dialect that is
equivalent to a mov with dpp specified, which is used to lower
gpu.warp_swizzle on AMD devices.