Add miopen.in_warp_transpose, its lowering, and a standalone test
This creates an op, in_warp_transpose that takes a vector of N
values, which are spread over T threads, and transposes each SxS block
within the NxT matrix logically formed by those values, where S is 2
or 4 (for now), optionally permuting the result rows (which thread
gets what) of the transpose. This operation will be used to enable
more efficient stores (and, eventually, loads) in the convolutions we
generate.