[MLIR][GPU] Define gpu.printf op and its lowering
- Define a gpu.printf op, which can be lowered to any GPU printf()
support (which is present in CUDA, HIP, and OpenCL). This ope only
supports constant format strings and scalar arguments
- Define the lowering of gpu.pirntf to a call to printf() (which is
what is required for AMD GPUs when using OpenCL) as well as to the
hostcall interface present in the AMD Open Compute device library,
which is the interface present when kernels are running under HIP.
- Add a "runtime" enum that allows specifying which of the possible
runtimes a ROCDL kernel will be executed under or that the runtime is
unknows. This enum controls how gpu.printf is lowered

This change does not enable lowering for Nvidia GPUs, but such a
lowering should be possible in principle.