Outline tosa::Conv2DOp and adjacent elementwise ops into kernel functions
Gather a convolution op and the leading and trailing element-wise ops
that feed into it and feed from it, and make a separate kernel function
from them.  Recognise identical kernels and merge them.  Implements the
--tosa-partition-pipeline option.

Includes unit test for the partition and trivial runnable test that
compares partitioned against non-partitioned against golden results.