Simplify index diff map computation
1. Constant fold more of the computation if possible.
This both simplifies the generated IR and enables further
optimizations, such as

2. Refactor carry application in Merge/Unfold operations.
Instead of a bunch of if statements, use remainders and integer
divisions to perform the carry logic. While in more complex cases,
such as the padding-1 kernel in the PR E2E tests, this ultimately
optimizes to the same ISA (at least, looking at the store logic),
it has the advantage of being correct when the overflow in a merge in
more than 1. In addition, this patch removed a lot of store address
computation logic from the stores in

-fil_layout=kyxc -in_layout=nhwc -out_layout=nhwk -batchsize=1 -in_channels=64 -in_h=2 -in_w=2 -out_channels=64 -fil_h=1 -fil_w=1 --dilation_h=1 --dilation_w=1 --conv_stride_h=1 --conv_stride_w=1 --padding_h=0 --padding_w=0 -t f32 --operation conv2d -x2 -pv -c

where the generated ISA now has a bunch of buffer_store_dword
instructions with constant offsets, which demonstrates the potential
value of this change.