Patches to internal files.
The biggest component is changes needed to fit dialect updates -- the
arith dialect was split out, there's a bufferization dialect, and so on.
Those affect both test files and source files.  A special bit of fun is
that both mlir::arith::ConstantOp and mlir::ConstantOp exist, and are
not the same.

Another big change is that affine maps associated with operations are no
longer an array of maps, but a single map, which implies that
composition happens as the map is set up rather than later.

Finally, I fiddled with libraries for libLLVMROCmBackendUtils.so, hoping
to avoid link failures.  I saw many "liblldDriver.so not found" failures.