Make the merging of identical kernels part of the outlining process, (#458)
rather than a separate pass after the partition pass.

That makes the separate pass, and the follow-up symbolDCE pass,
unnecessary, and thus makes the partition-pipeline pass also
unnecessary.  It uses some of the logic of mergeIdenticalBlocks() but
different code, since it isn't manipulating Blocks.

Change --tosa-partition from function-level to module-level pass,
because it creates FuncOps.  There were possibilities for race
conditions there.  In conjunction, move candidates list from module- to
function-level to avoid some trouble with shared nodes.

The one catch is that we no longer catch identical kernel functions that
originate in different primary functions, since we have only a
function-level view instead of a module-level view when merging.