Simplify XDlopsGemmV2 to use buffer offsets instead of transforms
The only reason XdlopsGemmV2 participates in the coordinate
transformation machinery is so that it can learn the offset at which
the LDS buffers for A and B start. XdlopsGemmV2 needs these offsets so
that we can allocate one LDS buffer instead of two (note to reviewer:
why?).

It's much simpler to pass in these offsets directly. This also allows
us to improve the quality of the generated code by moving certain
computations outside of the load/store loops.