Refactor how gemm IDs are computed for backward data convolution.
Introduce a routine to populate the list of gemm_id needed for a backward data
convolution, with the option to compute if a zero initialization kernel is
needed.

In case a zero initialization kernel is needed, its gemm_id would be -1.

Change the lowering pipeline to emit the zero initialization kernel when the
gemm_id is -1 in a backward data convolution.