Integrate OOB from transforms code
- Keep old version of backward data store kernel in order to
ensure that doesn't break for now. Removing it is a followup PR.

We now separately compute left and right OOB dims.

Compared to the initial proposal, handling of embed() is different to
account for backwards data.

This commit doesn't yet fix up the dialect tests or add tests for the
oob dims computation itself.