Fix bounds check rules for unfold
The bounds checks rules for unfold were causing valid writes to be
dropped in some backwards weight kernels. Having considered the
semantics of unfold, I've determined that, just like with merge,
checking the leftmost coordinate is correct, since an out of bounds
input to the unfold() will cause the leftmost coordinate of the unfold
to overflow.