Use index diffs everywhere (#582)
The lack of index diff use on padding kernels was partly due to an
llvm issue that has been fixed. Another part of the problem was that
removing the loadStoreHack caused certain wrw kernels to fail in the
non-xdlops case. What I think happened there is that unfold() didn't
properly propagate index updates that would have required writes to be
dropped. For example, if we had c = 255, y = x = 3 and we started at
(gemmN = 254 * 9 + 2 * 3 + 2) -> (254, 2, 2), using index diff maps to
compute gemmN + 1 would give (254, 2, 3), which would pass the oob
checker and be written, though the correct coordinate, (255, 0, 0)
would be oob.

Therefore, we disable unfold() in certain padding kernels, allowing us
to enable index diffs everywhere.

Fixes
https://github.com/ROCmSoftwarePlatform/llvm-project-private/issues/440