Fix gemm size calculation, probably fixing some bwd data bugs
The value for leftPadW was actually the value for rightPadH

This fixes the two configs specifically called out in
https://github.com/ROCmSoftwarePlatform/llvm-project-private/issues/534
but I'll need to re-test to confirm if that is the entire issue.

For the second suspicious config (with strides=2 in both h and w)
we get a change in grid_size from 72 to 96 on the kernels, which
matches the failure state (which was that no one was writing to
input[48:64]).