Define a shell script for building our Dockerfile
Currently, the Dockerfile build is inline in a Jenkins job config,
which makes it impossible to track changes and. Move the logic from
that build pipeline (with minor refactoring) to a script that the
Jenkins job can call.

This PR will have no effect until the rocm-docker job on ml-ci.amd.com
is changed to use the new script.