Move the computation of the thread offset in LDS to blockwise_gemm
In order to make progress on removing hardcoded math in gridwise_gemm,
moves the computation of thread offsets into a map that lives in
blockwise_gemm. This means that gridwise_gemm no longer needs to
compute the offset in the m and n dimensions that each particular
thread should use.

In addition
- Rename the level0 and level1 clusters to be more descriptive. The
level 0 cluster represented the 16 threads of a CU wave being arranged
into a mPerCuwave x nPerCuwave grid, and the level 1 cluster was a 2D
grid made of the cuwaves in a block. More descriptive names will
hopefully save future refactorers the headache of understanding what
all this cluster stuff is about
- Change the IR format of blockwise_gemm in pursuit of the general
aesthetics project