Remove dead code for dest_data_per_write
Review of tuning parameters shows that the write to LDS is always
tuned to a scalar write. Therefore, all the code in
gridwise-gemm-to-blockwise designed to handle the non-scalar case is
dead and has been for a quite a while. So, remove the tuning
parameter and all the code that depends on it.

Also, get rid of a test that's been doing nothing for years.