Remove dead PaddingInfoAttr
Since we now use bounds inference to determine when bounds checks are
needed, and since the bugs that prevented us from using index diffs in
padding kernels are long gone, there are no uses of the
PaddingInfoAttr in any of the gemm implementations ... but we still
pass it in.

Remove the attribute, remove the code setting it, and rework any
relevant tests to account for the loss of the information about
padding being applied.

This is also setup for a miopen.gemm operation.