Resolve incorrect results by handling alignment in the autovectorizer
Previously, if we had, for example, an image with width 4 padded to
width 6 (on the right) with a filter of 2, the vectorizer would infer
a vectorization length of 2 in the M dimension because, all else being
held constant, you could fetch the vectors (I[0], I[1]) (I[2],
I[3]), [faild bounds check -> (0, 0)] without partial oob.

However, suppose we had added 1 to the input coordinate to handle the
case where x = 1. Then we'd fetch the vectors (I[1], I[2]), (I[3],
I[4]), (0, 0). Note that I[4], which is reading into the next row and
should've been (I[3], 0) instead.

This commit resolves the issue by tracking alignment in the
vectorizer, and using that information to resolve the vectorization of
padding.