Move the cuwave related parameters out of tuning parameters attribute
Instead of deriving the {m,n}{ThreadsPerCuwave,CuwavesPerBlock}
attributes during tuning and packing them into the tuning parameters
struct when they're not actually tuning parameters, use a function to
compute those parameters as needed based on the block size. The new
function is identical to the old logic in the tuning parameter pass.

This has the advantage of making the non-xdlops tuning parameters
attribute exactly equivalent to the tuning parameter string we get
from external clients (possibly mod ordering).