More precise clamping workarounds for the vectorizer
Because the old vectorization logic makes some assumptions about the
relationship between the vectorization dimensions and kPack that don't
need to be true, it can fail if we don't properly clamp the results of
the vectorizer.

These clamps should be removed entirely when moving to the new
load/store loop structure.