[NFC] Make rock.transform only take one map
Having rock.transform take an array of maps was an archaic thing we
don't use anymore that just causes complications and unforseen edge
cases. That support was there from back when we didn't use
untransform() or transforming_for but instead had the transformations
as attributes on things like threadwise_copy_v2. This caused a lot of
headache and was removed quite a while back.

Fixes
https://github.com/ROCmSoftwarePlatform/llvm-project-private/issues/676