Create RockGemmWrapperInterface and use it in a few spots.
This is the first part of cleaning up the tensor_conv2d hack.
We'll now have an interface for gemm-like things, which we can extend
in the future (so that, for instance, AffixTuningParameters doesn't
need to be a template).

The getGemmSize() method is a refactoring, but getOutArgument() is
important for implementing bufferization generically over all the
conv2d* and gemm operations in a followup commit.