Recject padding kernels with kPack > 1 in a more principled way
Currently, we were rejecting padding + kPack > 1 by manually setting
kPack to 1 at the end of the tuning parameter inference. This can
cause problems by caksing (kPerBlock * mPerBlock * kPack) / waveSize
to be zero in gridwise gemm, spooking the vectorizer.

So, instead of that change, we just reject "needing to apply pading"
when there's kPack.

I expect this can go away eventually once we don't have the old
global load math, but that's the future.