Replacing tuning decided vectorization into transformMap based one (#768)
* Replacing tuning decided vectorization into transform based one

* Addressing review feedbacks

* Fixing typo

* Fixing a/b global copy size

* Enforce source data per read to follow kpack as upper bound

* More precise clamping workarounds for the vectorizer

Because the old vectorization logic makes some assumptions about the
relationship between the vectorization dimensions and kPack that don't
need to be true, it can fail if we don't properly clamp the results of
the vectorizer.

These clamps should be removed entirely when moving to the new
load/store loop structure.

* Make the buffer_load a global_load to handle wide vectors

Also, fix misc. compiler warnings

* Fix some compiler warnings

* More clamping to fix division by 0 problem

* Recject padding kernels with kPack > 1 in a more principled way

Currently, we were rejecting padding + kPack > 1 by manually setting
kPack to 1 at the end of the tuning parameter inference. This can
cause problems by caksing (kPerBlock * mPerBlock * kPack) / waveSize
to be zero in gridwise gemm, spooking the vectorizer.

So, instead of that change, we just reject "needing to apply pading"
when there's kPack.

I expect this can go away eventually once we don't have the old
global load math, but that's the future.

* Forgot to clang-format

* Rejecting perf config where block size is too large

* Fixing clamp caused derive to zero

* Fixing clang-format issues

Co-authored-by: Krzysztof Drewniak <Krzysztof.Drewniak@amd.com>