Rework system devices.
The system devices are no longer a StringMap of chip -> device,
because that's not sufficient to know if a device is usable. Instead,
they now have a vector of pointers into the device lists that were
produced by each detector.

And now there is a find() method that takes a partial architecture
name and reports if we have it on the system.

In addition, features are now stored for all targets as a
StringMap<bool> so we don't have to deal with the varying string
semantics as much.

Also, be sure to sort ROCm features alphabetically to keep things
canonical.