Use a smaller rebuild in CMake, saving CI time
Instead of clearing the build cache to switch shared/static lib, set
the CMake file up to FORCE the relevant settings (shared libraries, PR
tests, and so on) ON or OFF when someone explicitly passes in
-DBUILD_FAT_LIBROCKCOMPILER . Note that we introduce a variable that
tracks when the forcing has taken place to ensure that we don't FORCE
over options when we don't need to - that is, if we build static
library and then build the normal build, we need to force the options
back, but not in any other case, where we might've wanted to set those
options manually.

I expect this will save on CI time.

Fixes
https://github.com/ROCmSoftwarePlatform/llvm-project-private/issues/689