[NFC] Rename threadwise_copy_v2 to glabal_store
Change some names in AlignTiling.cpp to make it slightly clearer
what's going on.

Also, amend the fusion flow from BufferLoad to GlobalLoad, which is
one level higher up the stack (it can be one or more BufferLoad ops)
so as to prevent any potential issues when the output loop vectorizes
to a global_store that involves multiple buffer_store operations.