Fix cases where too much vectorization was applied in fusion
There exist some fusions (such as broadcasting a scalar) where we
could get bad results where the vectorization of the gemm writeback is
larger than the vectorization that the fusion argument (which is
transformed to give it the same shape as the gemm output).

Address this bug by having the load loop for additional arguments be
vectorized by a vectorization that's a function of the fusion-argument
transforms and capped at the vectorization of the writeback. This will
almost certainly lead to scalar loads.

Furthermore, resolve a long-standing bug in index diff updates that
could lead to use-before-def but somehow didn't trigger until this bug fix.