Add pass to fold equivalent buffer load operations (#851)
Because, in our kernels, there is only one variable that refers to a
given input buffer and is passed to the buffer load operations (that
is, while a fusion arg might be expand_shape'd, we never use both the
unexpanded and expanded memrefs), and because all buffer arguments are
distinct pointers on pain of us glaring angrily at you, we can cheaply
do an optimization that's harder prove in general: merging memory
loads

that is, if we have
x = load A[i]
...
y = load A[i]

, we can replace all uses of y with x and delete the second load.

Long term, if LLVM knew that buffer loads were loads, we could rely on
the infrastructure there to handle this more carefully. Or,
equivalently, if MLIR had a better alias analysis, we could take
advantage of it to be more cautions with it.

As it is, this code is in its own pass (to make it easier to check if
that's the issue) and is safe given our current and future plans for
our kernel generator, so I think we can risk putting something like
this in.