Clean up gemm_id, removing it or renaming it to kernelId
In backward data, we need a modified form of the kernel ID in order to
identify which section of the result that particular kernel invocation
is responsible for. This used to be handled by an implicit gemm_id
attribute and is now handled by an explicitly specified kernelId
attribute.

gemm_id also used to be used to identify utility kernels in the
backwards weight (or backwards data) case, but this is no longer the
case. Therefore, for all non-backwards data kernels, remove the
gemm_id attribute entirely.

Also, edit tests to reflect the new reality.

And, while I was here, make all the arch attributes in tests be fully
specified to match the behavior of the convolution generator.