Fix and begin testing mixed-layout convolutions
Fixes
https://github.com/ROCmSoftwarePlatform/llvm-project-private/issues/734

- Adds a transformation that runs before conv to gemm conversion that
ensures that the filter and input agree on the order of what will be
the K dimension (NHW/CYX) and that the input and output agree on the
order of N, H, and W. Not having this pass causes incorrect results as
the dimensions can be merged in the wrong order
- This pass uses the input tensor as the source of the true layout
because I had to pick one and because, in the forward case, it has
both gemmK and gemmN dimensions that need to match the output
- Fix the lowering_wrw_atomic_transform_nhwc test to actually be NHWC
- Add unit tests for the layout regularization transforms
- Add E2E tests that check various mixed-layout cases using a slightly
larger version of the config that was the minimal reproducer for the
earlier bug.
- As a result of said E2E tests, correct the assumption in forward and
wrw kernels that N always comes before HW
- Note: the backwarrd data kernels are currently unaffected since they
implicitly transform the kernel inputs to NCHW in the process of
generating the implicit GEMM

I've manually run these new tests on gfx1030, but I'll want a nightly
shared libs run to make sure I didn't miss anything.