Make elementwise utility kernels their own operations (#856)
* Make elementwise utility kernels their own operations

Fixes
https://github.com/ROCmSoftwarePlatform/llvm-project-private/issues/730

This gets rid of the gemm_id hack for making """conv2d*""" operations
that were actually elementwise kernels (ex. zero init, copying
workspace to filter).

- Define rock.zero_init_kernel and rock.converting_copy_kernel to
represent the two elementwise kernels that we need for some backwards
weight and backwards data cases
- Remove the gemm_id hack from --rock-conv-to-gemm and replace it with
rewrite patterns for the elementwise kernels
- Use Conv2DGenerator to select the appropriate utility kernel ops
when asked to generate the "convolution kernel" corresponding to one
of the utility calls
- Fix up the utility kernel rewrite patterns to allow for kernels that
don't vectorize properly - we probably haven't been hitting this
usecase but we also have no solid reason to make that sort of
thing (as we could get from, for example, bacwards data with a stride
of 3) inappricable
- Change the "arch" attribute on Conv2DGenerator results to be the
full arch name for consistency with everything else
- Update tests to match

* Clean up gemm_id, removing it or renaming it to kernelId

In backward data, we need a modified form of the kernel ID in order to
identify which section of the result that particular kernel invocation
is responsible for. This used to be handled by an implicit gemm_id
attribute and is now handled by an explicitly specified kernelId
attribute.

gemm_id also used to be used to identify utility kernels in the
backwards weight (or backwards data) case, but this is no longer the
case. Therefore, for all non-backwards data kernels, remove the
gemm_id attribute entirely.

Also, edit tests to reflect the new reality.

And, while I was here, make all the arch attributes in tests be fully
specified to match the behavior of the convolution generator.

* Update utility kernel vectorization length to take elemsPerThread.