Remove a bunch of old applicability/tuning restrictions
Going off of the comments and/or what they do, I can find no
fundamental reason that the restrictions removed (or, in the case of
kPack for various datatypes, loosened) needed to exist now that
Jerry's refactoring of gridwise_gemm_v2 (and other previous coordinate
transforms refactorings) need to exist.

This commit message will be updated with explanations for fixes for
any bugs that this removal uncovers.