Convert conv2d_wrw_perf_config tests into toml file
Since they are guarded my mfma=on only, we now use a config_name.cfg
file to keep track of any additional guards for the tests in the
config_name. This is better since we do not need to check arch info in
cmake configuration time, but rather during lit configuration time.