Use llvm.assume to encode bounds on block and grid sizes
Since our rock kernels know their launch dimensions, we can bound the
workgroup and workitem IDs from above by those dimensions. LLVM has an
assume intrinsic that is, in effect, an assert() that doesn't actually
generate code. The manual does warn that you shouldn't use this to
annotate thing the optimizer could infer, but the launch dimensions of
a kernel are not such a piece of information.

Performance measurements for Navi21 and comparisons of the generated
assembly will be added as comments to the PR once I have them.