Add a script to facilitate adding and searching configs from the rock E2E database
This PR adds a script, i.e. `rockE2E.sh`, to keep track of config in
the ROCK E2E framework. It can search configs, provided
as *.mlir, *.toml, or command line options, from one or more *.toml
files, which is considered as a conv2d config database.
The script can also insert the config into the toml file if the toml
file does not contain the config.

Use case: in the future when we want to add more configs into the ROCK
E2E framework, this script can help us find out whether the configs
are already included.