Remove Tosa layerwise constant folding to avoid mixed layout in Rock
We had this pass added early in our Tosa pipeline for reasons I'm not
entirely clear on. The tosa to linalg passes will run this
transformation, and those passes run after tosa-to-rock, and so
therefore we don't lose the fact that constant tensors are NCHW. We
also won't have to go through a lot of dramatics to un-transpose the
constants.